+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - [Docs] Fixed typo in Alerta docs with incorrect number of seconds in a day. - @jertel
 - Update GitHub actions to avoid running publish workflows on forked branches. - @jertel
 - Rewrite `_find_es_dict_by_key` per [discussion #1450](https://github.com/jertel/elastalert2/discussions/1450) for fieldnames literally ending in `.keyword` [#1459](https://github.com/jertel/elastalert2/pull/1459) - @jmacdone @jertel
+- [Opsgenie] Fix KeyError handling in OpsGenieAlerter custom message formatting to provide a default fallback message if a key is missing. - @mrsymlove
 
 # 2.18.0